Introduced CakePHP-like Model classes.
Introduced Model.class.php.
Model class is a basic O/R mapper,
and will load data from tables, and handles
associations of each models.

For example,

  $study = new Study($study_instance_uid);
  print($study->study_instance_uid);

will create a new instance of 'Study', which will have
field data retrieved from the 'study_list' table.

  $patient = $study->Patient;

print $patient->patient_name;
This will automatically retrieve the Patient data
which this study belongs to.

  $series = $study->Series;

This will automatically retrieve the list of Series
data which belong to this study.

The associations between tables are defined in static fields
of each Model subclass. Available association types
are 'hasMany', 'belongsTo', and 'hasAndBelongsToMany'.
The definition resembles CakePHP, but is not compatible.

Currently, searching and table joining is not implemented,
so these framework can be used in simple tasks, but works
good in CAD display page.

cad_result.php and some templates were modified
to use this framework.