Combine several javascript files and do some refactoring.
'jq-btn.js', 'hover.js' and 'radio-to-button.js' files are
combined to one file 'circus-common.js'. And reduced unnecessary codes.

Stylization is automatically done when DOM is ready, as usual,
but if you need stylization after the document is loaded,
for example by AJAX loading, you can call 'autoStylize()' method,
which was defined as extension of jQuery.