Introduce 'Job' class and rewrite ExecutePluginAction.
The new mode class 'Job' is created.
This Job class is a model class that contains the logic
for job insertion.
Simplified ExecutePluginAction class, which serves
as a simple controller class, and just calls
the Job::registerNewJob() method.

Also fixed the bug in detecting duplicate job.
This bug caused the same job (same plugin, same combination
of series) to be inserted repeatedly.

fixes #105